NEW: useDepositForm util and hook usage
